--- @types/interfaces.d.ts
@@ -1,3 +1,0 @@
-import { SelectElement } from '../src/vaadin-select.js';
-
-export type SelectRenderer = (root: HTMLElement, select?: SelectElement) => void;
