+++ vaadin-select.d.ts
@@ -1,2 +1,1 @@
 export * from './src/vaadin-select.js';
-export * from './@types/interfaces';
